:recycle: Extract locale extraction logic into helpers.rs
- Add extract_locale helper function to reduce code duplication
- Remove ~10 lines of repeated code from 7 formatter files
