+++ ext/icu4x/src/collator.rs
@@ -1,6 +1,5 @@
 use crate::data_provider::DataProvider;
 use crate::helpers;
-use crate::locale::Locale;
 use icu::collator::Collator as IcuCollator;
 use icu::collator::CollatorPreferences;
 use icu::collator::options::{CaseLevel, CollatorOptions, Strength};
@@ -79,19 +78,7 @@ impl Collator {
     /// * `case_first:` - :upper, :lower, or nil (default)
     fn new(ruby: &Ruby, args: &[Value]) -> Result<Self, Error> {
         // Parse arguments: (locale, **kwargs)
-        if args.is_empty() {
-            return Err(Error::new(
-                ruby.exception_arg_error(),
-                "wrong number of arguments (given 0, expected 1+)",
-            ));
-        }
-
-        // Get the locale
-        let locale: &Locale = TryConvert::try_convert(args[0])?;
-        let locale_ref = locale.inner.borrow();
-        let locale_str = locale_ref.to_string();
-        let icu_locale = locale_ref.clone();
-        drop(locale_ref);
+        let (icu_locale, locale_str) = helpers::extract_locale(ruby, args)?;
 
         // Get kwargs (optional)
         let kwargs: RHash = if args.len() > 1 {
