+++ ext/icu4x/src/helpers.rs
@@ -1,4 +1,6 @@
-use magnus::{Error, RHash, RModule, Ruby, Value, prelude::*};
+use crate::locale::Locale;
+use icu_locale::Locale as IcuLocale;
+use magnus::{Error, RHash, RModule, Ruby, TryConvert, Value, prelude::*};
 
 /// Resolves the provider from kwargs or falls back to the default provider.
 ///
@@ -23,4 +25,31 @@ pub fn resolve_provider(ruby: &Ruby, kwargs: &RHash) -> Result<Value, Error> {
             Ok(default)
         }
     }
+}
+
+/// Extracts and validates the locale from variadic arguments.
+///
+/// # Arguments
+/// * `ruby` - The Ruby runtime reference
+/// * `args` - The variadic arguments passed to the Ruby method
+///
+/// # Returns
+/// A tuple of (IcuLocale, String) where String is the locale's string representation.
+///
+/// # Errors
+/// Returns an error if no arguments are provided or if the first argument
+/// is not a valid Locale.
+pub fn extract_locale(ruby: &Ruby, args: &[Value]) -> Result<(IcuLocale, String), Error> {
+    if args.is_empty() {
+        return Err(Error::new(
+            ruby.exception_arg_error(),
+            "wrong number of arguments (given 0, expected 1+)",
+        ));
+    }
+    let locale: &Locale = TryConvert::try_convert(args[0])?;
+    let locale_ref = locale.inner.borrow();
+    let locale_str = locale_ref.to_string();
+    let icu_locale = locale_ref.clone();
+    drop(locale_ref);
+    Ok((icu_locale, locale_str))
 }
