+++ ext/icu4x/src/plural_rules.rs
@@ -1,6 +1,5 @@
 use crate::data_provider::DataProvider;
 use crate::helpers;
-use crate::locale::Locale;
 use fixed_decimal::Decimal;
 use icu::plurals::{
     PluralCategory, PluralRuleType, PluralRules as IcuPluralRules, PluralRulesPreferences,
@@ -31,20 +30,7 @@ impl PluralRules {
     /// * `type:` - :cardinal (default) or :ordinal
     fn new(ruby: &Ruby, args: &[Value]) -> Result<Self, Error> {
         // Parse arguments: (locale, **kwargs)
-        if args.is_empty() {
-            return Err(Error::new(
-                ruby.exception_arg_error(),
-                "wrong number of arguments (given 0, expected 1+)",
-            ));
-        }
-
-        // Get the locale
-        let locale: &Locale = TryConvert::try_convert(args[0])?;
-        let locale_ref = locale.inner.borrow();
-        let locale_str = locale_ref.to_string();
-        // Clone the locale before dropping the borrow
-        let icu_locale = locale_ref.clone();
-        drop(locale_ref);
+        let (icu_locale, locale_str) = helpers::extract_locale(ruby, args)?;
 
         // Convert to PluralRulesPreferences
         let prefs: PluralRulesPreferences = (&icu_locale).into();
